eu-hotel-scanner-v0.9/eu-hotel-scanner-v0.9/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "eu-hotel-scanner"
version = "0.9.0"
description = "EU hotel price scanner with cost index, multi-vendor aggregation, optimiser, and API service."
readme = "README.md"
requires-python = ">=3.9"
authors = [
  { name = "ChatGPT", email = "noreply@example.com" }
]
dependencies = [
  "PyYAML>=6.0",
  "requests>=2.31",
  "streamlit>=1.32",
  "python-dotenv>=1.0",
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
]

[project.scripts]
eu-hotel-scan = "hotel_scanner.cli:main"
eu-hotel-api = "service.api:run"

[tool.setuptools.packages.find]
where = ["."]
include = ["hotel_scanner*", "service*"]
